Removed compiler warnings on gnuc
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@715 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/tests/test_feature/source/testcdecl_class.cpp
@@ -104,7 +104,7 @@ bool TestCDecl_Class()
 
 	if( c1.a != 0xDEADC0DE )
 	{
-		printf("%s: Failed to assign object returned from function. c1.a = %X\n", TESTNAME, c1.a);
+		printf("%s: Failed to assign object returned from function. c1.a = %X\n", TESTNAME, (unsigned int)c1.a);
 		fail = true;
 	}
 
@@ -121,13 +121,13 @@ bool TestCDecl_Class()
 
 	if( c2.a != 0xDEADC0DE )
 	{
-		printf("%s: Failed to assign object returned from function. c2.a = %X\n", TESTNAME, c2.a);
+		printf("%s: Failed to assign object returned from function. c2.a = %X\n", TESTNAME, (unsigned int)c2.a);
 		fail = true;
 	}
 
 	if( c2.b != 0x01234567 )
 	{
-		printf("%s: Failed to assign object returned from function. c2.b = %X\n", TESTNAME, c2.b);
+		printf("%s: Failed to assign object returned from function. c2.b = %X\n", TESTNAME, (unsigned int)c2.b);
 		fail = true;
 	}
 
@@ -144,19 +144,19 @@ bool TestCDecl_Class()
 
 	if( c3.a != 0xDEADC0DE )
 	{
-		printf("%s: Failed to assign object returned from function. c3.a = %X\n", TESTNAME, c3.a);
+		printf("%s: Failed to assign object returned from function. c3.a = %X\n", TESTNAME, (unsigned int)c3.a);
 		fail = true;
 	}
 
 	if( c3.b != 0x01234567 )
 	{
-		printf("%s: Failed to assign object returned from function. c3.b = %X\n", TESTNAME, c3.b);
+		printf("%s: Failed to assign object returned from function. c3.b = %X\n", TESTNAME, (unsigned int)c3.b);
 		fail = true;
 	}
 
 	if( c3.c != 0x89ABCDEF )
 	{
-		printf("%s: Failed to assign object returned from function. c3.c = %X\n", TESTNAME, c3.c);
+		printf("%s: Failed to assign object returned from function. c3.c = %X\n", TESTNAME, (unsigned int)c3.c);
 		fail = true;
 	}
